Deployment on windows platform - pip install pyinstaller - pyinstaller --onefile main.py
After running this command, PyInstaller will create a dist/ directory in the same directory as your script. Inside dist/, you will find your executable file named main (on macOS and Linux) or main.exe (on Windows). This executable doesn't require a Python interpreter or any dependencies to run, which means you can run this on a system that doesn't have Python installed.

However, the executable is platform-specific, meaning an executable built on Windows will only run on Windows, and similarly for macOS and Linux. If you want your application to run on all platforms, you would have to build it separately on each platform.